+++ relay-server/src/endpoints/attachments.rs
@@ -62,7 +62,7 @@ pub async fn handle(
     let envelope = extract_envelope(meta, path, multipart, state.config()).await?;
     common::handle_envelope(&state, envelope)
         .await?
-        .ignore_rate_limits();
+        .check_rate_limits()?;
     Ok(StatusCode::CREATED)
 }